Adding "tag" as optional configuration parameter
Enable repo+tag to be explicitly set in conf.yml. When set, old
prefixing behavior applies. When absent, fallback on original behavior.

Change only impacts image name, container names left unchanged.

The other change you will notice is that I have changed the orchestrator
calls used to (a) get image id and (b) determine if an image exists to
use the "inspect" API instead of the "find". The filtering ceased to
work as soon as an image name contained a colon i.e. no results
returned. The undesirable implication of this is that the existence
check relies on catching an exception.